move components to folders
