+++ src/components/product/product.js
@@ -1,9 +1,9 @@
-import counter from '../hocs/counter';
+import counter from '../../hocs/counter';
 
 import style from './product.module.css';
 
-import { ReactComponent as Minus } from '../icons/minus.svg';
-import { ReactComponent as Plus } from '../icons/plus.svg';
+import { ReactComponent as Minus } from '../../icons/minus.svg';
+import { ReactComponent as Plus } from '../../icons/plus.svg';
 
 function Product(props) {
   const { amount, decrement, increment } = props;
